+++ src/errors.ts
@@ -10,3 +10,9 @@ export class MoreThanOneDistinctOrAllError extends Error{}
 export class MoreThanOneWhereStepError extends Error{}
 
 export class InvalidConditionError extends Error{}
+
+export class DeleteWithoutConditionError extends Error{}
+
+export class InvalidLimitValueError extends Error{}
+
+export class InvalidOffsetValueError extends Error{}