+++ src/option.ts
@@ -6,6 +6,7 @@ export type BuilderOption = {
   addPublicSchemaName?: 'always'|'never'|'when other schema mentioned'
   addTableName?: 'always'|'when two tables or more' //TODO: add more options like 'when needed' and 'when conflict'
   addAsBeforeTableAlias?: 'always'|'never'
+  throwErrorIfDeleteHasNoCondition?: boolean
 }
 
 export type BuilderOptionRequired = Required<BuilderOption>
@@ -18,7 +19,8 @@ const defaultOption: BuilderOptionRequired = {
   addPublicSchemaName: 'never',
   addTableName: 'when two tables or more',
   addAsBeforeTableAlias: 'always',
-}
+  throwErrorIfDeleteHasNoCondition: true,
+} as const
 Object.freeze(defaultOption)
 
 export function fillUndefinedOptionsWithDefault(option: BuilderOption): BuilderOptionRequired {
@@ -30,5 +32,6 @@ export function fillUndefinedOptionsWithDefault(option: BuilderOption): BuilderO
     addPublicSchemaName: option.addPublicSchemaName ?? defaultOption.addPublicSchemaName,
     addTableName: option.addTableName ?? defaultOption.addTableName,
     addAsBeforeTableAlias: option.addAsBeforeTableAlias ?? defaultOption.addAsBeforeTableAlias,
+    throwErrorIfDeleteHasNoCondition: option.throwErrorIfDeleteHasNoCondition ?? defaultOption.throwErrorIfDeleteHasNoCondition,
   }
 }